refactor: organize file-transfer tests into tests/ subdirectory
- Move test files to src/file-transfer/tests/
- Move __screenshots__ directory to tests/
- Update imports to use relative paths (../)
- All 68 tests passing after reorganization
